REFACTOR: Move Cat submodules to top of `CategoricalAlgebra`.
A step toward making `Categories` and `FinCats` the foundation of
`CategoricalAlgebra`, as one would expect.
