+++ src/categorical_algebra/CategoricalAlgebra.jl
@@ -2,6 +2,8 @@ module CategoricalAlgebra
 
 using Reexport
 
+include("Categories.jl")
+include("FinCats.jl")
 include("FreeDiagrams.jl")
 include("Limits.jl")
 include("Subobjects.jl")
@@ -9,8 +11,6 @@ include("Sets.jl")
 include("FinSets.jl")
 include("Matrices.jl")
 include("FinRelations.jl")
-include("Categories.jl")
-include("FinCats.jl")
 include("CSets.jl")
 include("GraphCategories.jl")
 include("Diagrams.jl")
@@ -20,19 +20,19 @@ include("DataMigration.jl")
 include("StructuredCospans.jl")
 include("DPO.jl")
 
+@reexport using .Categories
+@reexport using .FinCats
 @reexport using .FreeDiagrams
 @reexport using .Limits
 @reexport using .Subobjects
 
 @reexport using .Sets
 @reexport using .FinSets
-@reexport using .Categories
-@reexport using .FinCats
 @reexport using .CSets
+@reexport using .CatElements
 
 @reexport using .Diagrams
 @reexport using .CommutativeDiagrams
-@reexport using .CatElements
 @reexport using .DataMigration
 @reexport using .StructuredCospans
 @reexport using .DPO
