+++ src/categorical_algebra/Categories.jl
@@ -13,14 +13,13 @@ instances are supported through the wrapper type [`TypeCat`](@ref). Finitely
 presented categories are provided by another module, [`FinCats`](@ref).
 """
 module Categories
-export Cat, TypeCat, Ob, Functor, Transformation,
-  dom, codom, compose, id, ob, hom, is_hom_equal,
-  ob_map, hom_map, dom_ob, codom_ob, component
+export Cat, TypeCat, Functor, Transformation, dom, codom, compose, id,
+  ob, hom, is_hom_equal, ob_map, hom_map, dom_ob, codom_ob, component
 
 using AutoHashEquals
 
-using ...GAT, ..Sets
-import ...Theories: Category2, Ob, ob, hom, dom, codom, compose, ⋅, ∘, id,
+using ...GAT
+import ...Theories: Category2, ob, hom, dom, codom, compose, ⋅, ∘, id,
   composeH, *
 
 # Categories
@@ -85,8 +84,6 @@ struct TypeCat{Ob,Hom} <: Cat{Ob,Hom} end
 
 TypeCat(Ob::Type, Hom::Type) = TypeCat{Ob,Hom}()
 
-Ob(::TypeCat{T}) where T = TypeSet{T}()
-
 # FIXME: This isn't practical because types are often too tight.
 #ob(::TypeCat{Ob,Hom}, x) where {Ob,Hom} = convert(Ob, x)
 #hom(::TypeCat{Ob,Hom}, f) where {Ob,Hom} = convert(Hom, f)
@@ -112,12 +109,6 @@ abstract type Functor{Dom<:Cat,Codom<:Cat} end
 """ Evaluate functor on morphism.
 """
 @inline hom_map(F::Functor, f) = do_hom_map(F, f)
-
-""" Forgetful functor Ob: Cat → Set.
-
-Sends a category to its set of objects and a functor to its object map.
-"""
-function Ob end
 
 @auto_hash_equals struct IdentityFunctor{Dom<:Cat} <: Functor{Dom,Dom}
   dom::Dom
