+++ src/categorical_algebra/Limits.jl
@@ -16,12 +16,11 @@ export AbstractLimit, AbstractColimit, Limit, Colimit,
 
 using AutoHashEquals
 
-using ...Theories
+using ...GAT, ...Theories
 import ...Theories: ob, terminal, product, proj1, proj2, equalizer, incl,
   initial, coproduct, coproj1, coproj2, coequalizer, proj,
   delete, create, pair, copair, factorize
 using ..FreeDiagrams
-using ...GAT
 import ..FreeDiagrams: apex, legs
 
 # Data types for limits
