+++ test/categorical_algebra/FinSets.jl
@@ -8,7 +8,6 @@ sshow(args...) = sprint(show, args...)
 # Finite sets
 #############
 
-# Finite ordinals.
 set = FinSet(3)
 @test length(set) == 3
 @test collect(set) == 1:3
@@ -26,6 +25,22 @@ set = FinSet(Set(1:2:5))
 set = FinSet((x=[1,3,5], y=['a','b','c']))
 @test length(set) == 3
 @test map(NamedTuple, set) == [(x=1, y='a'), (x=3, y='b'), (x=5, y='c')]
+
+# Discrete categories
+#####################
+
+C = FinCat(FinSet(3))
+@test C isa FinCat{Int,Int}
+@test is_discrete(C)
+@test collect(ob_generators(C)) == 1:3
+@test isempty(hom_generators(C))
+@test (dom(C, 1), codom(C, 1)) == (1, 1)
+@test (id(C, 2), compose(C, 2, 2)) == (2, 2)
+
+F = FinDomFunctor([FinSet(1), FinSet(3), FinSet(1)],
+                  C, TypeCat(FinSet{Int}, FinFunction{Int}))
+@test ob_map(F, 2) == FinSet(3)
+@test hom_map(F, 2) == id(FinSet(3))
 
 # Functions between finite sets
 ###############################
